@@ -9,7 +9,8 @@
   "dependencies": {
     "core-js": "^3.6.5",
     "socket.io": "^4.4.1",
-    "vue": "^2.6.11"
+    "vue": "^3.2.26",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
