@@ -1,8 +1,21 @@
-import Vue from 'vue'
-import App from './App.vue'
+import App from './App';
+import { createApp } from 'vue';
+import { createStore } from 'vuex';
+import io from 'socket.io-client';
+const socket = io("http://localhost:3000");
 
-Vue.config.productionTip = false
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+const store = createStore({
+  state () {
+    return {
+      socket
+    }
+  },
+  
+})
+
+
+const app = createApp(App);
+
+app.use(store);
+app.mount('#app');
