@@ -2,13 +2,13 @@
   <Login v-if="login" @submit="novoUsuario" />
   <JetChat v-else />
   
+  
 </template>
 
 <script>
 import Login from './components/Login.vue';
 import JetChat from  './components/JetChat.vue';
-import { io } from "socket.io-client";
-const socket = io("http://localhost:3000");
+
 
 export default {
   name: 'App',
@@ -25,20 +25,24 @@ export default {
     }
   },
   created(){
-    socket.emit("teste", "bom dia");
+    this.$store.state.socket.on('alreadyExist', user=>{
+      alert('Este usuario ja existe' + user.nome);
+      this.login = true;
+    })
   },
   methods:{
-    novoUsuario(e){
+       novoUsuario(e){
       e.preventDefault();
-      let nome = document.querySelector("input#nome");
-      if(nome.value != ''){
+      let nome = document.querySelector("input#nome").value.trim();
+      if(nome != '' ){
         let user = {
-          nome: nome.value.trim(),
-          id: socket.id,
+          nome: nome,
+          id: this.$store.state.socket.id
         }
         this.login = !this.login;
         this.user = user;
-        socket.emit('newUser', user);
+        this.$store.state.socket.emit('newUser', user)
+        
         
       }
       
