@@ -20,10 +20,15 @@ export default {
   
   data(){
     return{
-    
+      
     }
+  },
+  methods:{
+    
   }
+ 
 }
+
 </script>
 
 <style lang="scss">
