@@ -3,15 +3,21 @@
       <div class="side-bar">
           <h3>Users</h3>
           <ul>
-              <li v-for="(user, index) in users" :key="index">{{user}}</li>
+              <li v-for="(user, index) in users" :key="index">{{user.nome}}</li>
           </ul>
       </div>
       <div class="chat-area">
           <h1>JetChat</h1>
+          <div class="text-area">
+              <ul>
+                  <li v-for="(msg, index) in listMessage" :key="index">{{msg.author}} : {{msg.text}}</li>
+              </ul>
+          </div>
           <div class="message-area">
-              <input type="text" name="message" id="message">
+              
+              <input type="text" name="message" id="message" autocomplete="off" @keypress.enter="sendMessage">
               <i class="far fa-smile emoji"></i>
-              <i class="far fa-paper-plane send"></i>
+              <i class="far fa-paper-plane send" @click="sendMessage"></i>
           </div>
           
       </div>
@@ -26,17 +32,29 @@ export default {
     data(){
         return{
             users :[],
+            listMessage: []
         }
     },
     created(){
-        socket.on('listUsers', users=>{
-            console.log("bom dia");
-            this.users = users;
-            
-        })
+       this.$store.state.socket.on('listUsers', users=>{
+           this.users = users;
+       })
+       this.$store.state.socket.on('updateUsers', users=>{
+           this.users = users;
+       })
+       this.$store.state.socket.on('listMessages', messages =>{
+           this.listMessage = messages;
+
+       })
     },
     methods:{
-        
+        sendMessage(){
+            let inputMsg = document.querySelector('#message');
+            if(inputMsg.value != ''){
+                this.$store.state.socket.emit('newMessage', inputMsg.value);
+                inputMsg.value = '';
+            }
+        }
     }
 }
 </script>
@@ -77,6 +95,13 @@ export default {
                 margin: 10px;
                 text-transform: uppercase;
             }
+            .text-area{
+                height: calc(100% - 150px);
+                width: 100%;
+                ul{
+                    overflow-y:auto;
+                }
+            }
             .message-area{
                 width: 100%;
                 position:absolute;
@@ -91,6 +116,7 @@ export default {
                 }
                 .emoji, .send{
                     font-size: 30px;
+                    cursor: pointer;
                    
                 }
             }
